chore: 네이밍 컨벤션 준수 및 폴더 구조 변경 JUL-36
