+++ components/post/index.ts
